Basic wrapping of observables and computeds using object accessor descriptors.
The get accessor either propogates the access onto the observable or returns the observable depending on whether it is run in the context of an expose callback.

The set accessor always propagates the access onto the observable but changes the knockout semantics by explicitly returning the value and not this. However, I don't think returning this would be a usefull concept when using the assignment operator. (Actually the return value is implicit so we could remove this line.)

Currently there's no support for observableArrays.

This would break if you tried to combine it with a call to applyBindings, although it may be ok if the applyBindings call was itself wrapped in an expose callback. Will test this next...